% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{backTransform}
\alias{backTransform}
\title{Back transform to the simplex}
\usage{
backTransform(gcData, nPCs, transData, combinedChains)
}
\arguments{
\item{gcData}{List containing the geochemical and related data. This container is
described in the package documentation.}

\item{nPCs}{Number of principal components that were used in the finite mixture model.}

\item{transData}{List containing the transformed geochemical
concentrations and related information.
This list is return by function \code{\link{transformGcData}}, for which the
documentation includes a complete
description of container \code{transData}.}

\item{combinedChains}{A \code{stanfit} object containing multiple Monte Carlo chains. This
object is return by function \code{\link{combineChains}}, for which the
documentation includes a complete description of container
\code{combinedChains}.}
}
\value{
A list with the following components is returned.

\item{compMean1}{Matrix containing Monte Carlo
samples of the compositional mean vector for pdf 1.
The first dimension of the matrix
pertains to indices of the samples; the second dimension pertains to
indices of the compositional mean vector.}

\item{compMean2}{Matrix containing Monte Carlo
samples of the mean vector for pdf 2. Its structure is identical to
that for \code{compMean1}.}

\item{varMatrix1}{Array containing Monte Carlo samples of the
variation matrix
for pdf 1. The first dimension of the array pertains to indices of
the samples; the second and the third dimensions pertain to indices of
the variation matrix.}

\item{varMatrix2}{Array containing samples of the variation matrix
for pdf 2. Its structure is identical to that for \code{varMatrix1}.}
}
\description{
Back transform the mean vectors and covariance matrices
of the finite mixture model to the corresponding quantities for the simplex
(that is, compositional means and variation matrices).
}
\examples{
\dontrun{
simplexModPar <- backTransform( gcData, nPCs, transData, combinedChains)
}

}
\references{
Hron, K., Filzmoser, P., 2015, Exploring compositional data with the robust
compositional biplot, in Carpita, M., Brentari, E., Qannari, M., eds.,
Advances in latent variables - Methods, models and applications,
Springer, p 219-226.

Pawlowsky-Glahn, V., Egozcue, J.J., and Tolosana-Delgado, R., 2015, Modeling
and analysis of compositional data: John Wiley and Sons, Ltd.
}

