% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{calcCondProbs1}
\alias{calcCondProbs1}
\title{Calculate conditional probabilities}
\usage{
calcCondProbs1(transData, nPCs, combinedChains)
}
\arguments{
\item{transData}{List containing the transformed geochemical
concentrations and related information.
This list is return by function \code{\link{transformGcData}}, for which the
documentation includes a complete description of container \code{transData}.}

\item{nPCs}{Number of principal components that were used in the finite mixture model.}

\item{combinedChains}{A \code{stanfit} object containing multiple Monte Carlo chains. This
object is return by function \code{\link{combineChains}}, for which the
documentation includes a complete description of container
\code{combinedChains}.}
}
\value{
A matrix containing the Monte Carlo samples of the
conditional probabilites. The number of matrix columns equals the
number of field samples, and the number of matrix rows equals the
number of Monte Carlo samples in \code{combinedChains}. That is,
column j of the matrix contains Monte Carlo samples of the conditional
probability that field sample j is associated with the first probability
density function in the finite mixture model.
}
\description{
Calculate, for each field sample, Monte Carlo samples of
the conditional probability that the field sample is
associated with the first probability density function in the finite
mixture model.
}
\details{
The conditional probabilities are calculated using the formula in Gelman
et al. (2014, p. 540 bottom).
}
\examples{
\dontrun{
condProbs1 <- calcCondProbs1(transData, nPCs, combinedChains)
}


}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A.,
and Rubin, D.B., 2014, Bayesian data analysis (3rd ed.),
CRC Press.
}

