% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{calcObsTestStats}
\alias{calcObsTestStats}
\title{Calculate observed test statistics for model checking}
\usage{
calcObsTestStats(transData, nPCs, condProbs1)
}
\arguments{
\item{transData}{List containing the transformed geochemical
concentrations and related information.
This list is return by function \code{\link{transformGcData}}, for which the
documentation includes a complete description of container \code{transData}.}

\item{nPCs}{Number of principal components that were used in the
finite mixture model.}

\item{condProbs1}{A matrix containing the Monte Carlo samples of the
conditional probabilites. This matrix is returned by function
\code{\link{calcCondProbs1}}, for which the documentation includes a
complete description of container \code{condProbs1}.}
}
\value{
A list with the following test quantities is returned.

\item{mu1}{Mean vector for pdf 1.}

\item{mu2}{Mean vector for pdf 2.}

\item{tau1}{Standard deviation vector for pdf 1.}

\item{tau2}{Standard deviation vector for pdf 2.}

\item{Corr1}{Correlation matrix for pdf 1.}

\item{Corr2}{Correlation matrix for pdf 2.}
}
\description{
Calculate test statistics for posterior
predictive checking
of the parameters in the finite mixture model. The test statistics pertain
to the observed data that have undergone both the isometric log-ratio
transform and the robust, principal component transform.
}
\details{
Test statistics are defined in Gelman et al. (2014, p. 145) and
are used for posterior predictive checking of the finite mixture model.
}
\examples{
\dontrun{
obsTestStats <- calcObsTestStats(transData, nPCs, condProbs1)
}


}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B.,
Vehtari, A., and Rubin, D.B., 2014, Bayesian data analysis (3rd ed.):
CRC Press.
}

