% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{calcSimplexStats}
\alias{calcSimplexStats}
\title{Calculate sample statistics pertinent to the simplex}
\usage{
calcSimplexStats(gcData)
}
\arguments{
\item{gcData}{List containing the geochemical and related data. This container is
described in the package documentation.}
}
\value{
A list with three elements is returned. Variable D is the number
of geochemical concentrations reported for each sample.

\item{sampleCenter}{Vector of length D.}

\item{variationMatrix}{Matrix of dimension D x D.}

\item{metricVariance}{Scalar. The metric variance is also called
the total sample variance.}
}
\description{
Calculate the sample center, total variation matrix, and
the metric variance.
}
\details{
The sample statistics are described in
section 5.2 of Pawlowsky-Glahn et al. (2015).
}
\examples{
\dontrun{
simplexStats <- calcSimplexStats(gcData)
}

}
\references{
Pawlowsky-Glahn, V., Egozcue, J.J., and Tolosana-Delgado, R., 2015, Modeling
and analysis of compositional data: John Wiley and Sons, Ltd.
}

