% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{combineChains}
\alias{combineChains}
\title{Combine selected chains}
\usage{
combineChains(samplePars, selectedChains, procDir = ".")
}
\arguments{
\item{samplePars}{List containing, among other things, the names of the files in which the
\code{stanfit} objects are stored. These objects contain the samples
of the posterior pdf. This list is return by function
\code{\link{sampleFmm}}, for which the documentation includes
a complete description of container \code{samplePars}.}

\item{selectedChains}{Dataframe listing the indices of the chains that
are combined. (See Details).}

\item{procDir}{Directory containing the files with the \code{stanfit} objects.}
}
\value{
The returned value is a \code{stanfit} object,
which is described in the rstan documentation. This object comprises
the multiple chains, which have samples of
the following model parameters:
\tabular{ll}{
Parameter \tab Description \cr
theta     \tab Proportion of population associated with pdf 1.\cr
mu1       \tab Mean vector for pdf 1.\cr
mu2       \tab Mean vector for pdf 2.\cr
tau1      \tab Standard deviation vector for pdf 1.\cr
tau2      \tab Standard deviation vector for pdf 2.\cr
L_Omega1  \tab Cholesky decomposition of the correlation matrix for pdf 1.\cr
L_Omega2  \tab Cholesky decomposition of the correlation matrix for pdf 2.\cr
log_lik   \tab The logarithm of the likelihood function.
}
}
\description{
Combine and arrange selected, individual chains to create a set
of chains.
}
\details{
Argument \code{selectedChains} is a dataframe with two columns. The first
column, for which the heading is "Chain," comprises the indices of the chains
that will be combined. The second column, for which the heading is
"isSwitched", comprises logical values (that is, TRUE or FALSE), which
indicate whether the variables in the associated chain are switched.

To understand \code{selectedChains} consider two examples. (1) Assume that
row 1 of \code{selectedChains} comprises the values 2 and FALSE. These
values
indicate that chain 2 will be included in the combination and that
the variables in the
finite mixture model are not switched. That is, variable mu1 in the
chain is assigned to mu1, variable mu2 in the chain is assigned to mu2, and
so on. (2) Assume that
row 2 of \code{selectedChains} comprises the values 4 and TRUE. These
values
indicate that chain 4 will be included in the combination and
that the variables in the
finite mixture model are switched. That is, variable mu1 in the
chain is assigned to mu2, variable mu2 in the chain is assigned to mu1, and
so on.
}
\examples{
\dontrun{
combinedChains <- combineChains(samplePars, selectedChains)
}

}
\references{
Stan Development Team, 2015, Stan Modeling Language - Users Guide
and Reference Manual, Version 2.6.0, available on line at
\url{http://mc-stan.org/} (last accessed October 2015).
}

