% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{plotCompMeans}
\alias{plotCompMeans}
\title{Plot compositional means}
\usage{
plotCompMeans(simplexModPar, elementOrder, symbolSize = 2, symbolShape = 1,
  yLimits = NULL, isEePlotted = FALSE)
}
\arguments{
\item{simplexModPar}{List containing Monte Carlo samples of the selected parameters
in the finite mixture model.
These parameters (namely, the mean vector, the standard deviation vector,
 and the correlation matrix for
each pdf) are expressed in terms of their equivalent values in the
simplex (namely, the compositional mean vector and the variation matrix
for each pdf).
This list is return by function \code{\link{backTransform}}, for which the
documentation includes a complete description of this container.}

\item{elementOrder}{Vector specifying the order in which the elements are plotted.}

\item{symbolSize}{The size of the plotting symbol.}

\item{symbolShape}{The shape of the plotting symbol.}

\item{yLimits}{Vector of length 2 specifying the limits of the y-axis. If an element is NA
then ggplot computes the corresponding axis limit.}

\item{isEePlotted}{Logical variable specifying whether the amalgamated concentration "EE",
which accounts for all omitted and unmeasured element concentrations, is
included in the plot.}
}
\description{
Plot the compositional means for the pdfs in the
finite mixture model.
}
\details{
The compositional mean is a vector. The Monte Carlo samples of each v
ector element are summarized by their median.
}
\examples{
\dontrun{
plotCompMeans( simplexModPar, gcData, elementOrder)
}

}

