% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{plotEdaCorr}
\alias{plotEdaCorr}
\title{Plot the correlations among the principal components}
\usage{
plotEdaCorr(transData)
}
\arguments{
\item{transData}{List containing the transformed geochemical
concentrations and related information.
This list is return by function \code{\link{transformGcData}}, for which the
documentation includes a complete description of container \code{transData}.}
}
\description{
Plot both the correlation matrix of the principal components and
the histogram of the correlations.
}
\examples{
\dontrun{
plotEdaCorr(transData)
}

}

