% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{plotEdaDist}
\alias{plotEdaDist}
\title{Plot the principal component distributions}
\usage{
plotEdaDist(transData)
}
\arguments{
\item{transData}{List containing the transformed geochemical
concentrations and related information.
This list is return by function \code{\link{transformGcData}}, for which the
documentation includes a complete description of container \code{transData}.}
}
\description{
Plot the distributions of the principal components as
both boxplots and violinplots.
}
\examples{
\dontrun{
plotEdaDist(transData)
}

}

