% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{plotEdaVar}
\alias{plotEdaVar}
\title{Plot the scree plot}
\usage{
plotEdaVar(transData, relOffset = 0.04, size = 3)
}
\arguments{
\item{transData}{List containing the transformed geochemical
concentrations and related information.
This list is return by function \code{\link{transformGcData}}, for which the
documentation includes a complete description of container \code{transData}.}

\item{relOffset}{Scalar specifying the relative distance that each
percentage of cumulative variance (see Details) is offset from the top
of its respective bar.}

\item{size}{Scalar specifying the text size for the percentage of
cumulative variance (see Details).}
}
\description{
Plot the variances associated with the (robust) principal
components.
}
\details{
In principal component analysis, this plot is called a "scree plot."

The variance for each component is represented by a bar, and above each
bar is the cumulative percentage of total variance. The total variance
is the sum of the variances for each component. The percentage of variance
associated a component is its variance divided by the total variance
times 100. The cumulative percentage of total variance for component k
is the sum of the percentages from component 1 to component k.
}
\examples{
\dontrun{
plotEdaVar(transData)
}

}
\references{
Johnson, R.A., and Wichern, D.W., 2007, Applied multivariate statistical
analysis (6th ed.): Pearson Prentice Hall.
}

