% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{plotPointStats}
\alias{plotPointStats}
\title{Plot point statistics for each chain}
\usage{
plotPointStats(samplePars, procDir = ".", excludedChains = NULL)
}
\arguments{
\item{samplePars}{List containing, among other things, the names of the files in which the
\code{stanfit} objects are stored. These \code{stanfit} objects contain
the samples
of the posterior pdf, which are needed to calculate the point statistics.
List \code{samplePars} is return by function
\code{\link{sampleFmm}}, for which the documentation includes
a complete description of the list.}

\item{procDir}{Directory containing the files with the \code{stanfit} objects.}

\item{excludedChains}{Vector with the indices of the chains for which
the point statistics are not used to calculate plot ranges. See Details.}
}
\description{
Plot point statistics for each chain. The point statistics are
calculated and plotted for element 1 of the
two mean vectors, element 1 of the two standard deviation vectors,
the proportion associated with pdf1, and the log-likelihood.
}
\details{
To prevent a point statistic for a particular Monte Carlo sampling chain
from being included in the calculation of
the plot range, the index of that chain is specified in
argument \code{excludedChains}. In other words, the point statistics for
all chains are calculated and plotted. If a few chains are specified by
argument \code{excludedChains}, then the calculation of the plot ranges
does not use those specified chains.

The calculation of the point statistics for every chain requires a
minute or two.
}
\examples{
\dontrun{
plotPointStats(samplePars)
}

}

