% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{plotSelectedTraces}
\alias{plotSelectedTraces}
\title{Plot selected traces}
\usage{
plotSelectedTraces(samplePars, procDir = ".")
}
\arguments{
\item{samplePars}{List containing, among other things, the names of the files in which the
\code{stanfit} objects are stored. Each \code{stanfit} object contains
one chain comprising samples
of the posterior pdf.
List \code{samplePars} is return by function
\code{\link{sampleFmm}}, for which the documentation includes
a complete description of the list.}

\item{procDir}{Directory containing the files with the \code{stanfit} objects.}
}
\description{
Plot selected traces for each chain. Analysis of these traces
shows whether within-chain label switching has occurred and whether the
traces are stationary.
}
\details{
Three plots are generated for each chain:
\itemize{
 \item Trace of the proportion associated with pdf 1
 \item Traces of the first element of the two mean vectors.
 \item Traces of the first element of the two standard deviation
 vectors.
}
}
\examples{
\dontrun{
plotTraces(samplePars, procDir)
}

}

