% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{plotSqrtVarMatrices}
\alias{plotSqrtVarMatrices}
\title{Plot the combined, scaled variation matrices}
\usage{
plotSqrtVarMatrices(simplexModPar, elementOrder, colorScale = "spectrum",
  scaleRange = NULL, isEePlotted = FALSE)
}
\arguments{
\item{simplexModPar}{List containing Monte Carlo samples of four parameters
in the finite mixture model.
These parameters (namely, the mean vector and covariance matrix for
each pdf) are expressed in terms of their equivalent values in the
simplex (namely, the compositional mean vector and the variation matrix
for each pdf).
This list is return by function \code{\link{backTransform}}, for which the
documentation includes a complete description of this container.}

\item{elementOrder}{Vector specifying the order in which the elements are plotted.}

\item{colorScale}{Character string specifying the color scale for
the plot. The choices are either "spectrum" (default) and "rainbow."}

\item{scaleRange}{Vector of length 2 specifying the range of the color scale.}

\item{isEePlotted}{Logical variable specifying whether the amalgamated concentration "EE",
which accounts for all omitted and unmeasured element concentrations, is
included in the plot.}
}
\description{
Plot the combined, scaled variation matrices.
Because a variation matrix is symmetric and its diagonal is zero, the
two variation matrices from the two pdfs are combined into a single matrix.
The matrix is then scaled by the square root function, to reduce its range.
}
\details{
In the plot, the upper triangle is the upper triangle from the
variation matrix for pdf 1, and the lower triangle is the lower triangle
from the variation matrix for pdf 2.

The pixels represent scaled variances of the log-ratios
between the respective chemical elements. The scaling is desirable because
it reduces the large range of the variances, making it easier to
visualize all of the variances together. The scaling function is the
square root; so, the pixels represent standard deviations
of the log-ratios between the respective chemical elements.
}
\examples{
\dontrun{
plotSqrtVarMatrix(simplexModPar, elementOrder, colorScale = "rainbow")
}

}
\references{
Pawlowsky-Glahn, V., Egozcue, J.J., and Tolosana-Delgado, R., 2015, Modeling
and analysis of compositional data: John Wiley and Sons, Ltd.
}

