% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{sampleFmm}
\alias{sampleFmm}
\title{Sample the posterior pdf}
\usage{
sampleFmm(transData, nPCs, sm, priorParams, nWuSamples = 500,
  nPwuSamples = 500, nChainsPerCore = 2, nCores = parallel::detectCores(),
  procDir = ".")
}
\arguments{
\item{transData}{List containing the transformed geochemical
concentrations and related information.
This list is return by function \code{\link{transformGcData}}, for which the
documentation includes a complete description of container \code{transData}.}

\item{nPCs}{Number of principal components that are used in the
finite mixture model (See Details).}

\item{sm}{S4 class stanmodel containing the compiled, stan model.}

\item{priorParams}{Parameters for the prior pdfs (See Details)}

\item{nWuSamples}{Number of warm-up samples in each chain (See details).}

\item{nPwuSamples}{Number of post warm-up samples in each chain (See details).}

\item{nChainsPerCore}{Number of chains that each core computes (See Details).}

\item{nCores}{Number of central processing unit (cpu) cores that are
used (See Details).}

\item{procDir}{Directory into which the results are written
(See Details).}
}
\value{
The returned values may be conveniently divided into two groups.
The first group comprises the \code{stanfit} objects that contain the
sampling chains. There is one \code{stanfit} object for each chain; the
format of the \code{stanfit} object is described in the rstan documentation.
The samples within an object are of the following model parameters:
\tabular{ll}{
Parameter \tab Description \cr
theta     \tab Proportion of population associated with pdf 1.\cr
mu1       \tab Mean vector for pdf 1.\cr
mu2       \tab Mean vector for pdf 2.\cr
tau1      \tab Standard deviation vector for pdf 1.\cr
tau2      \tab Standard deviation vector for pdf 2.\cr
L_Omega1  \tab Cholesky decomposition of the correlation matrix for pdf 1.\cr
L_Omega2  \tab Cholesky decomposition of the correlation matrix for pdf 2.\cr
log_lik   \tab The logarithm of the likelihood function.
}
Each \code{stanfit} object is written to its own file in
directory \code{procDir}.
A file name has the form "RawSamples?-?.dat" for which the first
question mark is replaced by the core number and the the second question
mark replaced by the chain number for the specified core.
The important point is that the file name is unique.

The second group is a list with 4 elements that is returned by function
\code{sampleFmm}:

\item{nChains}{Number of chains.}

\item{nWuSamples}{Number of warm-up samples in each chain.}

\item{nPwuSamples}{Number of post warm-up samples in each chain.}

\item{fileNames}{Vector containing the names of the files
with the \code{stanfit} objects.}

The total number of samples per chain equals \code{nWuSamples} plus
\code{nPwuSamples}. Only the post warm-up samples are used for inference.
}
\description{
Sample the posterior probability density function for the
Bayesian formulation of the
finite mixture model. The model implementation and the sampling are
performed with package rstan.
}
\details{
The parameters in the finite mixture model are estimated
from the robust principal components, which are stored as a matrix within
\code{transData}. The number of principal components, \code{nPCs},
means that
principal components 1, 2, ..., nPCs are used in the finite mixture model.
That is, all principal components greater than nPCs are not used.

Argument priorParams is a vector with 4 elements.
\enumerate{
 \item{
The first element pertains to prior pdf for theta, which is the proportion
in the finite mixture model. The prior pdf is a beta pdf, which has two shape
parameters. The shape parameters must be equal so that the beta pdf is
symmetric with respect to 0.5. Also, the shape parameters must be greater than
1, so that the pdf has a peak at 0.5 and equals 0 at both 0 and 1. Both shape
parameters are specified by the single value, priorParams[1].
 }
 \item{
The second element pertains to the prior pdf for the elements of mu1 and
mu2, which are the mean vectors in the finite mixture model. The prior pdf
is a normal pdf, for which the mean is 0 and the standard deviation is
specified by priorParams[2]. Of course, priorParams[2] must be greater than 0.
 }
 \item{
The third element pertains to the prior pdf for the elements of tau1 and
tau2, which are the standard deviation vectors in the finite mixture model.
The prior pdf is a truncated Cauchy pdf: The Cauchy pdf before truncation has
a center of 0 and a scale specified by priorParams[3]. The truncation point is
0, which removes negative values of the random variable. Of course,
priorParams[3] must be greater than 0.
 }
 \item{
The fourth element pertains to the prior pdf for Omega1 and Omega2,
which are the correlation matrices in the finite mixture model.
The prior pdf is the LKJ correlation distribution, which has one shape
parameter. When the shape parameter is greater than 1, the LKJ correlation
distribution has a mode corresponding to the identity matrix; as the shape
parameter increases, the LKJ correlation distribution becomes increasingly
concentrated about this mode. The chosen shape parameter always should be
greater than 1. The shape parameter is specified by priorParams[4].
 }
}
Appropriate values for the elements of \code{priorParams} are selected
by examining the distribution of the principal components.

The sampling of the posterior probability density function (pdf)
has two parts: warm-up and post warmup. The number of samples
during warm-up is \code{nWuSamples}, and the number of samples during post
warm-up is \code{nPwuSamples}.

The sampling is performed in parallel on multiple central
processing unit (cpu) cores. Of course, the number of requested cores
\code{nCores} must be less than or equal to the actual number. The
parallel computations are organized in the following manner:
\code{nCores} cores are requested from the operating system; each core
computes \code{nChainsPerCore} individual chains, and each chain is written
to its own file in directory \code{procDir}.

The sampling method is Hamiltonian Monte Carlo sampling. If the sampling
has not converged using the default value for argument \code{nWuSamples},
then this argument should be increased. A suitable value might be 2500.
}
\examples{
\dontrun{
tmp <- normalizePath(path.package("GcClust"))
load(paste(tmp, "\\\\stan\\\\MixtureModel.bin", sep=""))

samplePars <- sampleFmm(transData, nPCs, sm, priorParams )
}

}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A.,
and Rubin, D.B., 2014, Bayesian data analysis (3rd ed.),
CRC Press.

Marin, J.M., Mengersen, K. and Robert, C.P., 2005,
Bayesian modelling and inference on mixtures of distributions,
in Handbook of Statistics 25, D. Dey and C.R. Rao (eds.),
Elsevier-Sciences.

Stan Development Team, 2015, Stan Modeling Language - Users Guide
and Reference Manual, Version 2.6.0, available on line at
\url{http://mc-stan.org/} (last accessed October 2015).
}

