% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GcClusterFunctions.R
\name{splitGcData}
\alias{splitGcData}
\title{Split the geochemical data}
\usage{
splitGcData(gcData, condProbs1, threshold = 0.1)
}
\arguments{
\item{gcData}{List containing the geochemical and related data. This container is
described in the package documentation.}

\item{condProbs1}{A matrix containing the Monte Carlo samples of the
conditional probabilites. This matrix is returned by function
\code{\link{calcCondProbs1}}, for which the documentation includes a
complete description of container \code{condProbs1}.}

\item{threshold}{The threshold used to split the data into two groups. (See details.)}
}
\value{
A list with four components is returned.

\item{gcData1}{List containing the geochemical and related data that
are associated with pdf 1. The format of this container is the same
as the format of \code{gcData} and is described in the package
documentation.}

\item{analysisData1}{Dataframe containing data that are helpful in
analyzing the samples associated with pdf 1. The data are the sample
number, the conditional probability associated with the sample, the
longitude, and the latitude.}

\item{gcData2}{List containing the geochemical and related data that
are associated with pdf 2. The format of this container is the same
as the format of \code{gcData} and is described in the package
documentation.}

\item{analysisData2}{Dataframe containing data that are helpful in
analyzing the samples associated with pdf 2. The data are the sample
number, the conditional probability associated with the sample, the
longitude, and the latitude.}
The information in components \code{analysisData1} and \code{analysisData2}
is also in components \code{gcData1} and \code{gcData2}, except for the
conditional probabilities. The reason for generating these containers with
redundant information is that their format is simple, so that the containers
can be passed to other programs for further analysis.
}
\description{
The geochemical data, which have been clustered, are split
into two groups based on their conditional probabilities.
}
\details{
For each
field sample, the median of the Monte Carlo samples of conditional
probability
is calculated. If this median is between 1-threshold and 1, then the
field sample is associated with pdf 1 in the finite mixture model. However,
if this median is between 0 and threshold, then the field sample is
associated with pdf 2 in the finite mixture model. This criterion is
used to split the field samples into two groups.

Argment \code{threshold} must be greater than 0 and less than 0.5.
}
\examples{
\dontrun{
theSplits <- splitGcData(gcData, condProb1, threshold = 0.10 )
}

}

